requirements.txt
fastapi
python-dotenv
uvicorn
PyPDF2
python-magic
